Add pass & fail cases to countries & header reducers with mock store
